+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- The `bit_string` module gains the `concat` function.
 - The `os` module has been moved to the `gleam_os` library.
 - The `rescue` function has been removed from the `function` library in favour
   of target specific versions in Erlang and JavaScript specific libraries.